Some refactoring done in the code: book request now have it's own controller
